requirements.txt
# Web Framework
flask==3.0.0
flask-login==0.6.3
werkzeug==3.0.1

# Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# NLP and AI
sentence-transformers==2.2.2
faiss-cpu==1.7.4
thefuzz==0.20.0

# Document Processing
pypdf==3.17.4
python-docx==1.1.0

# Mathematics Engine
sympy==1.12
scipy==1.11.4
matplotlib==3.8.2
plotly==5.17.0
scikit-learn==1.3.2
statsmodels==0.14.0

# Performance
numexpr==2.8.7

# Development
watchdog==3.0.0

# Production Server (for local testing)
gunicorn==21.2.0